Cleaned up the commentary in all of our controllers

+++ app/Http/Controllers/OrderController.php
@@ -17,7 +17,6 @@ use Illuminate\Validation\ValidationException;
 use Redirect;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
-use SearchHelper;
 
 
 class OrderController extends Controller
@@ -27,7 +26,7 @@ class OrderController extends Controller
      * Bookstore doesn't want us showing data that is too old so that profs are
      * less inclined to pick old books. Old books are often harder for the bookstore to acquire.
      */
-    const PAST_BOOKS_NUM_MONTHS = 24;
+    const PAST_BOOKS_NUM_MONTHS = 30;
 
 
     /** GET: /requests/
@@ -41,7 +40,12 @@ class OrderController extends Controller
     }
 
 
-    /** GET: /requests/create/{course_id?}
+    /** GET: /requests/create/{$course_id?}?book_id={book_id?}
+     *
+     * Displays the page for submitting textbook requests.
+     *
+     * If $course_id is provided, the user will be taken straight to book selection for that course.
+     * If book_id is set in the query string, the given book will already be in the cart once the user selects a course.
      *
      * @param Request $request
      * @param int|null $course_id
@@ -52,6 +56,9 @@ class OrderController extends Controller
         $user = $request->user();
         $openTerms = Term::currentTerms()->get();
 
+        // Maintain an array of all the parameters we will pass to our view.
+        // These here are the ones that are always the same.
+        // There are also some that are set below that depend on what path we take through this code.
         $viewParams = [
             'openTerms' => $openTerms,
             'current_user_id' => $user->user_id,
@@ -59,17 +66,24 @@ class OrderController extends Controller
         ];
 
         if ($course_id == null){
+            // In this path, the user will be presented with all courses that they are able to place orders for.
             $this->authorize('all');
 
             $openTermIds = $openTerms->pluck('term_id');
 
-            // If the user can place orders for everyone, don't return courses for everything here
-            // because it would almost certainly crash their browser if we did.
             if ($request->user()->may('place-all-orders')){
+                // If the user can place orders for everyone, don't return ALL courses here
+                // because it would almost certainly crash their browser if we did.
+                // This is only going to apply to bookstore staff and administrators.
+                // with a compsci rollout, Connie will still only have the place-dept-orders permission.
+
+                // Instead of showing all courses, we will just present the courses that they teach.
+                // Which probably isn't any courses, but there isn't really anything else to do here that makes sense.
                 $query = $request->user()->courses();
                 $viewParams['continueUrl'] = '/courses';
             }
             else{
+                // If the user can only place dept orders, or if they can only place their own orders, then get all orderable courses.
                 $query = Course::orderable();
             }
 
@@ -79,6 +93,7 @@ class OrderController extends Controller
                     'listings',
                     'orders.book',
                     'user' => function($query){
+                        // Only select what we need from users for security/information disclosure purposes.
                         return $query->select('user_id', 'first_name', 'last_name');
                     }])
                 ->get();
@@ -86,15 +101,16 @@ class OrderController extends Controller
             $viewParams['courses'] = $courses;
 
 
-            $book = [];
-            if ($request->input('book_id') != null) {
-                $book = Book::with('authors')
-                    ->where('book_id', '=', $request->input('book_id'))
-                    ->get();
+            // If a book_id was passed in to the query string of the request,
+            // send that along to the view so that it can be added to the cart initially.
+            if ($request->input('book_id')) {
+                $book = Book::with('authors')->findOrFail($request->input('book_id'));
+                $viewParams['book'] = $book;
             }
-			$viewParams['book'] = $book;
         }
         else {
+            // In this path, the user will be taken straight to a specific course.
+            /** @var Course $course */
             $course = Course::findOrFail($course_id);
 
             $this->authorize('place-order-for-course', $course);
@@ -109,6 +125,7 @@ class OrderController extends Controller
                     'listings',
                     'orders.book',
                     'user' => function($query){
+                        // Only select what we need from users for security/information disclosure purposes.
                         return $query->select('user_id', 'first_name', 'last_name');
                 }]);
 
@@ -128,7 +145,9 @@ class OrderController extends Controller
     }
 
 
-    /** POST: /requests/no-book/{order_id}
+    /** POST: /requests/no-book/
+     *
+     * Mark the given course as not needing a book. If the course has any outstanding orders, they will be deleted.
      *
      * @param Request $request
      * @return \Illuminate\Http\JsonResponse
@@ -141,9 +160,11 @@ class OrderController extends Controller
         $this->authorize('place-order-for-course', $course);
 
         foreach ($course->orders as $order) {
-            $order->deleted_by = $request->user()->user_id;
-            $order->save();
-            $order->delete();
+            if (!$order->deleted_by){
+                $order->deleted_by = $request->user()->user_id;
+                $order->save();
+                $order->delete();
+            }
         }
 
         $course->no_book = true;
@@ -155,6 +176,8 @@ class OrderController extends Controller
 
 
     /** POST: /requests/delete/{order_id}
+     *
+     * Deletes the specified order.
      *
      * @param Request $request
      * @param $order_id
@@ -176,11 +199,14 @@ class OrderController extends Controller
 
     /** GET: /requests/past-courses/{course_id}
      *
+     * Retrieves the past offerings of the specified course within the last PAST_BOOKS_NUM_MONTHS months.
+     *
      * @param $course_id int course id to get books for.
      * @return \Illuminate\Http\JsonResponse
      */
     public function getPastCourses($course_id)
     {
+        /** @var Course $course */
         $course = Course::findOrFail($course_id);
 
         $this->authorize('place-order-for-course', $course);
@@ -190,9 +216,11 @@ class OrderController extends Controller
         // This gives us any courses that are effectively the same course as this one.
         $similarCourseIdsQuery = $course->getSimilarCourseIdsQuery();
 
+        // We need to get the full course models of those similar course_ids,
+        // and we also need to restrict by age. The bookstore doesn't want super old
+        // past books showing up because old books are hard for them to get.
         $courses =
-            Course::
-            whereIn('course_id', $similarCourseIdsQuery)
+            Course::whereIn('course_id', $similarCourseIdsQuery)
             ->where('created_at', '>', Carbon::today()->subMonths(static::PAST_BOOKS_NUM_MONTHS))
             ->with([
                     "term" => function ($query){
@@ -204,6 +232,7 @@ class OrderController extends Controller
                             ->with('book.authors');
                     },
                     'user' => function($query){
+                        // Only select what we need from users for security/information disclosure purposes.
                         return $query->select('user_id', 'last_name');
                     }])
             ->get();
@@ -251,6 +280,8 @@ class OrderController extends Controller
             $courses[] = $course;
         }
 
+
+        // Save the results of each request so that we can echo feedback to the user.
         $orderResults = [];
 
 
@@ -260,7 +291,10 @@ class OrderController extends Controller
 
                 $book = $bookData['book'];
 
+                // If the book doesn't have a book_id, it must be a new book.
                 if (!isset($book['book_id']) || !$book['book_id']) {
+
+                    // Clean up the ISBN so that we can look for it in the database.
                     $isbn = isset($book['isbn13']) ? $book['isbn13'] : '';
                     $isbn = preg_replace('|[^0-9]|', '', $isbn);
                     $edition = trim(isset($book['edition']) ? $book['edition'] : '');
@@ -268,9 +302,13 @@ class OrderController extends Controller
                     $db_book = null;
 
                     if ($isbn) {
+                        // If the isbn was provided, look for an existing book in the database with the same ISBN.
                         $db_book = Book::where('isbn13', '=', $isbn)->first();
                     }
                     else {
+                        // If the isbn was not provided, then the user clicked the button to add a new
+                        // book to their cart that does not have an ISBN.
+                        // Check for an existing book with the same title, pub, and edition.
                         $db_book = Book::where([
                             'title' => trim($book['title']),
                             'publisher' => trim($book['publisher']),
@@ -279,6 +317,8 @@ class OrderController extends Controller
                     }
 
                     if (!$db_book) {
+                        // If the book was not already found in the database,
+                        // create a new book.
                         $db_book = Book::create([
                             'title' => trim($book['title']),
                             'isbn13' => $isbn,
@@ -294,20 +334,26 @@ class OrderController extends Controller
                     }
 
                 } else {
+                    // If the book_id was set on the book, it must be an
+                    // already-existing book. Go get it.
                     $db_book = Book::findOrFail($book['book_id']);
                 }
 
+                // Clean up the notes for the book.
                 $notes = isset($bookData['notes']) ? $bookData['notes'] : '';
                 $notes = trim($notes);
 
+                // Make a spot in our order results for this book if we haven't already made one.
                 $book_id = $db_book->book_id;
                 if (!isset($orderResults[$book_id]))
                     $orderResults[$book_id] = [];
 
-                $existingOrders = $course->orders;
-                $sameOrders = $existingOrders->where('book_id', $book_id);
+                // Check if there are already orders for this course for this book.
+                $sameOrders = $course->orders->where('book_id', $book_id);
 
                 if (count($sameOrders)) {
+                    // If there are already orders for this book for this course,
+                    // echo those back to the user.
                     foreach ($sameOrders as $sameOrder) {
                         $orderResults[$book_id][] = [
                             'notPlaced' => true,
@@ -320,7 +366,10 @@ class OrderController extends Controller
                         ];
                     }
                 }
-                else{
+                else
+                {
+                    // There are no other orders for this course for this book.
+                    // Create a new order in the database.
                     $order = $course->orders()->save(new Order([
                         'notes' => $notes,
                         'placed_by' => $user_id,
@@ -328,9 +377,12 @@ class OrderController extends Controller
                         'book_id' => $book_id
                     ]));
 
-                    // Reload orders so we can detect duplicate orders.
+                    // Reload orders so we can detect duplicate orders
+                    // in subsequent iterations of the loops we're in.
+                    // (This important in case the user picked what ended up being the same book more than once).
                     $course->load('orders');
 
+                    // Record this order as being successful so the user can be aware of what happened.
                     $orderResults[$book_id][] = [
                         'notPlaced' => false,
                         'order' => $order->load('book.authors'),
@@ -339,6 +391,8 @@ class OrderController extends Controller
                 }
             }
 
+            // Since we just created an order for this course, it can't possibly have no book.
+            // Clear out these fields to keep our data consistent.
             $course->no_book = false;
             $course->no_book_marked = null;
             $course->save();
